+++ lib/oauth/authentication-result.js
