+++ lib/oauth/authenticator.js
@@ -4,9 +4,9 @@ var njwt = require('njwt');
 
 var ApiAuthRequestError = require('../error/ApiAuthRequestError');
 
-var OAuthAuthenticationResult = require('./OAuthAuthenticationResult');
+var OAuthAuthenticationResult = require('./authentication-result');
 
-var OAuthPasswordGrantRequestAuthenticator = require('./OAuthPasswordGrantRequestAuthenticator');
+var OAuthPasswordGrantRequestAuthenticator = require('./password-grant').authenticator;
 
 function OAuthAuthenticator(application) {
   if (!(this instanceof OAuthAuthenticator)) {
