+++ lib/stormpath.js
@@ -5,6 +5,6 @@ module.exports = {
   loadApiKey: require('./authc/ApiKeyLoader'),
   Client: require('./Client'),
   Tenant: require('./resource/Tenant'),
-  OAuthPasswordGrantRequestAuthenticator: require('./authc/OAuthPasswordGrantRequestAuthenticator'),
-  OAuthAuthenticator: require('./authc/OAuthAuthenticator')
+  OAuthPasswordGrantRequestAuthenticator: require('./oauth/password-grant').authenticator,
+  OAuthAuthenticator: require('./oauth/authenticator')
 };
